[Task-11] Move fetch survey list to adapter
